+++ humility/temp/test_storybook/.storybook/main.js
@@ -1,7 +1,7 @@
 /** @type { import('@storybook/react-vite').StorybookConfig } */
 const config = {
   stories: ['../src/**/*.mdx', '../src/**/*.stories.@(js|jsx|mjs|ts|tsx)'],
-  addons: [],
+  addons: ['@storybook/addon-essentials', '@storybook/addon-interactions'],
 
   framework: {
     name: '@storybook/react-vite',